remove BookController interface.
