--- hw09-classic-web/src/main/java/ru/otus/atheneum/controller/BookControllerImpl.java
@@ -1,93 +1,0 @@
-package ru.otus.atheneum.controller;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.servlet.view.RedirectView;
-import ru.otus.atheneum.dto.AuthorRow;
-import ru.otus.atheneum.dto.BookRow;
-import ru.otus.atheneum.dto.GenreRow;
-import ru.otus.atheneum.service.AuthorService;
-import ru.otus.atheneum.service.BookService;
-import ru.otus.atheneum.service.EntityConverter;
-import ru.otus.atheneum.service.GenreService;
-import ru.otus.domain.Book;
-
-@Controller
-@Slf4j
-@RequiredArgsConstructor
-public class BookControllerImpl implements BookController {
-
-  private final BookService bookService;
-  private final AuthorService authorService;
-  private final GenreService genreService;
-  private final EntityConverter entityConverter;
-
-  @Override
-  @GetMapping("/")
-  public String bookList(Model model) {
-    List<BookRow> books = bookService.getAll().stream().map(entityConverter::convertBookEntityToDto)
-        .collect(Collectors.toList());
-    log.info(books.toString());
-    model.addAttribute("books", books);
-    return "books";
-  }
-
-  @Override
-  @GetMapping("/book/delete/{id}")
-  public RedirectView deleteBook(Model model, @PathVariable("id") String id) {
-    bookService.delete(id);
-    log.info(String.format("Удаление книги с id = %s", id));
-    return new RedirectView("/", true);
-  }
-
-  @Override
-  @GetMapping("/book/edit/{id}")
-  public String editBook(Model model, @PathVariable("id") String id) {
-    Book book = bookService.getById(id)
-        .orElseThrow(
-            () -> new BookControllerException(
-                String.format("Нет книги по переданному идентификатору %s", id)));
-    BookRow bookRow = entityConverter.convertBookEntityToDto(book);
-    prepareModelForEditMapping(model, bookRow);
-    return "book_edit";
-  }
-
-  @Override
-  @GetMapping("/book/edit")
-  public String createBook(Model model) {
-    BookRow bookRow = new BookRow();
-    prepareModelForEditMapping(model, bookRow);
-    return "book_edit";
-  }
-
-  @Override
-  @PostMapping("/book/edit")
-  public RedirectView savePerson(
-      BookRow bookRow,
-      Model model
-  ) {
-    log.info(bookRow.toString());
-    bookService.update(entityConverter.convertBookDtoToEntity(bookRow));
-    return new RedirectView("/", true);
-  }
-
-  private void prepareModelForEditMapping(Model model, BookRow bookRow) {
-    List<AuthorRow> authors = authorService.getAll().stream()
-        .map(entityConverter::convertAuthorEntityToDto)
-        .collect(Collectors.toList());
-    List<GenreRow> genres = genreService.getAll().stream()
-        .map(entityConverter::convertGenreEntityToDto)
-        .collect(Collectors.toList());
-    model.addAttribute("book", bookRow);
-    model.addAttribute("authors", authors);
-    model.addAttribute("genres", genres);
-  }
-
-}
